fix bug with image urls
Images with relative urls were being loaded.
That is because the files are mounted on public/files/.
Therefore, if the browser tries, for example, to request
the an image with url /a/img.png, it will request the
file public/a/img.png, but it should request the file
public/files/a/img.png. This issue has been fixed using
JavaScript to replace relative image urls with their right
url.